# docs/datastore_api_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "The echoSMs web API",
    "version": "0.1.0"
  },
  "paths": {
    "/v1/datasets": {
      "get": {
        "tags": [
          "get"
        ],
        "summary": "Get dataset_ids with optional filtering",
        "operationId": "get_datasets_v1_datasets_get",
        "parameters": [
          {
            "name": "species",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Species",
              "description": "The scientific species name"
            },
            "description": "The scientific species name"
          },
          {
            "name": "imaging_method",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Imaging method",
              "description": "The imaging method used"
            },
            "description": "The imaging method used"
          },
          {
            "name": "model_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model type",
              "description": "The model type used"
            },
            "description": "The model type used"
          },
          {
            "name": "aphiaID",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "AphiaID",
              "description": "The [aphiaID](https://www.marinespecies.org/aphia.php)"
            },
            "description": "The [aphiaID](https://www.marinespecies.org/aphia.php)"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of dataset_ids",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/dataset/{dataset_id}": {
      "get": {
        "tags": [
          "get"
        ],
        "summary": "Get the dataset with the given dataset_id",
        "operationId": "get_dataset_v1_dataset__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The dataset ID",
              "title": "Dataset Id"
            },
            "description": "The dataset ID"
          },
          {
            "name": "full_data",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "If true, all raw data for the dataset will be returned as a zipped file",
              "default": false,
              "title": "Full Data"
            },
            "description": "If true, all raw data for the dataset will be returned as a zipped file"
          }
        ],
        "responses": {
          "200": {
            "description": "A dataset structured as per the echoSMs data store [schema](https://ices-tools-dev.github.io/echoSMs/schema/data_store_schema/)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specimens/{dataset_id}": {
      "get": {
        "tags": [
          "get"
        ],
        "summary": "Get the specimen_ids from the dataset with the given dataset_id",
        "operationId": "get_specimens_v1_specimens__dataset_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The dataset ID",
              "title": "Dataset Id"
            },
            "description": "The dataset ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A list of specimen_ids",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specimen/{dataset_id}/{specimen_id}": {
      "get": {
        "tags": [
          "get"
        ],
        "summary": "Get specimen data with the given dataset_id and specimen_id",
        "operationId": "get_specimen_v1_specimen__dataset_id___specimen_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The dataset ID",
              "title": "Dataset Id"
            },
            "description": "The dataset ID"
          },
          {
            "name": "specimen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The specimen ID",
              "title": "Specimen Id"
            },
            "description": "The specimen ID"
          }
        ],
        "responses": {
          "200": {
            "description": "A specimen dataset structured as per the echoSMs data store [schema](https://ices-tools-dev.github.io/echoSMs/schema/data_store_schema/)",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/specimen_image/{dataset_id}/{specimen_id}": {
      "get": {
        "tags": [
          "get"
        ],
        "summary": "Get an image of the specimen shape, with the given dataset_id and specimen_id",
        "operationId": "get_specimen_image_v1_specimen_image__dataset_id___specimen_id__get",
        "parameters": [
          {
            "name": "dataset_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The dataset ID",
              "title": "Dataset Id"
            },
            "description": "The dataset ID"
          },
          {
            "name": "specimen_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "The specimen ID",
              "title": "Specimen Id"
            },
            "description": "The specimen ID"
          }
        ],
        "responses": {
          "200": {
            "description": "An image of the specimen shape",
            "content": {
              "image/png": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
  "tags": [
    {
      "name": "get",
      "description": "Calls that get information from the data store"
    }
  ]
}